+++ app/views/items/index.html.erb
@@ -125,32 +125,21 @@
   <h2 class='title'>ピックアップカテゴリー</h2>
   <div class="subtitle">新規投稿商品</div>
   <ul class='item-lists'>
-    <% if @items.present? %>
-      <% @items.each do |item| %>
-        <li class='list'>
-          <%= link_to item_path(item) do %>
-            <div class='item-img-content'>
-             <%= image_tag item.image if item.image.attached? %>
-              <%# <% if item.sold_out? %>
-                <%# <div class='sold-out'> %>
-                  <%# <span>Sold Out!!</span> %>
-                <%# </div> %>
-              <%# <% end %>
+    <li class='list'>
+      <%= link_to '#' do %>  <!-- ここはプレースホルダーでアイテムがない状態 -->
+        <%= image_tag "https://tech-master.s3.amazonaws.com/uploads/curriculums/images/Rails1-4/sample.jpg", class: "item-img" %>
+        <div class='item-info'>
+          <h3 class='item-name'>商品を出品してね！</h3>
+          <div class='item-price'>
+            <span>99999999円<br>(税込み)</span>
+            <div class='star-btn'>
+              <%= image_tag "star.png", class:"star-icon" %>
+              <span class='star-count'>0</span>
             </div>
-            <div class='item-info'>
-              <h3 class='item-name'><%= item.name %></h3>
-              <div class='item-price'>
-                <span><%= item.price %>円<br><%= item.shipping_fee %></span>
-                <div class='star-btn'>
-                  <%= image_tag "star.png", class:"star-icon" %>
-                  <span class='star-count'>0</span>
-                </div>
-              </div>
-            </div>
-          <% end %>
-        </li>
+          </div>
+        </div>
       <% end %>
-    <% else %>
+        </li>
       <li class='list'>
         <%= link_to '#' do %>
           <%= image_tag "https://tech-master.s3.amazonaws.com/uploads/curriculums/images/Rails1-4/sample.jpg", class: "item-img" %>
@@ -166,7 +155,6 @@
           </div>
         <% end %>
       </li>
-    <% end %>
   </ul>
 </div>
   <%# /商品一覧 %>
