+++ spec/models/item_spec.rb
@@ -19,6 +19,12 @@ RSpec.describe Item, type: :model do
         @item.valid?
         expect(@item.errors.full_messages).to include("User must exist")
       end
+
+      it 'imageが空では出品できない' do
+        @item.image = nil
+        @item.valid?
+        expect(@item.errors.full_messages).to include("Image can't be blank")
+      end
       
       it 'categoryが選択されていない場合出品できない' do
         @item.category_id = nil
@@ -39,27 +45,27 @@ RSpec.describe Item, type: :model do
       end
       
       it 'conditionが選択されていない場合出品できない' do
-        @item.condition_id = nil
+        @item.condition_id = 1
         @item.valid?
-        expect(@item.errors.full_messages).to include("Condition can't be blank")
+        expect(@item.errors.full_messages).to include("Condition must be other than 1")
       end
       
       it 'shipping_feeが選択されていない場合出品できない' do
-        @item.shipping_fee_id = nil
+        @item.shipping_fee_id = 1
         @item.valid?
-        expect(@item.errors.full_messages).to include("Shipping fee can't be blank")
+        expect(@item.errors.full_messages).to include("Shipping fee must be other than 1")
       end
       
       it 'prefectureが選択されていない場合出品できない' do
-        @item.prefecture_id = nil
+        @item.prefecture_id = 1
         @item.valid?
-        expect(@item.errors.full_messages).to include("Prefecture can't be blank")
+        expect(@item.errors.full_messages).to include("Prefecture must be other than 1")
       end
       
       it 'delivery_timeが選択されていない場合出品できない' do
-        @item.delivery_time_id = nil
+        @item.delivery_time_id = 1
         @item.valid?
-        expect(@item.errors.full_messages).to include("Delivery time can't be blank")
+        expect(@item.errors.full_messages).to include("Delivery time must be other than 1")
       end
       
       it 'priceが空では出品できない' do
@@ -78,6 +84,12 @@ RSpec.describe Item, type: :model do
         @item.price = 10000000
         @item.valid?
         expect(@item.errors.full_messages).to include("Price must be less than or equal to 9999999")
+      end
+
+      it 'priceが全角を含むと出品できない' do
+        @item.price = '３０００'
+        @item.valid?
+        expect(@item.errors.full_messages).to include("Price is not a number")
       end
 
   end
